Refactors ignorables
